02/22/2024 V4
EVERYTHING IS FIXED!!!!!!!!!!! IM SO SMART!!!!!!!!!!!!!!!!!!!!!!!!!!!!

implemented button (again)
need to rectify referentMeasure - why won't it work as a global variable?
********need to ask*********
need to rectify known error in which the pong ball will phase through the exit button.
edited if(){}else statements and rectified the order in which the fireworks were drawn so that they no longer spawn in front of the pong ball.
rectified known error in which the fireworks would spawn before touching the edge of the screen.