03/13/2024 V2
rectified an error in which the movedBall would not bounce off the paddles and would instead only obey physics if it was in a pleasant mood. It would occasionally also decide to only half-bounce and phase through the paddle before bouncing.

known error in which the fireworks also do not obey physics and ignore the existence of gravity. They just stay suspended in place. not sure why, but thats a tomorrow me issue.