03/21/2024
implemented a button instance
rectified an error in which font "roboto" would not load.
rectified an error in which the given font would not appear when drawn.
encapsulated text creation in the button constructor to preserve system resources.
optimized button hover over = rect(); is only drawn once instead of twice.