04/06/24
rectified a scoreboard error in which score value would update - but the button would not.
(moved button creation into draw .. probably not the most efficient, but this will be fixed and/or edited later on.)

changed idle button color to black and text color to white for a sleeker UI appearance. the confetti from PartyMode is created behind the buttons - but this is preferable to being drawn in front as the moving squares may block text and  overwhelm a user.  i intend to keep this the way it is.

implemented an error catch for the movedBall instance which rectified a known bug in which movedBall could be created outside of the table bounds.

implemented scoreUpdate() to rectifiy redunant code.

removed redundant comments from previous ball respawn function - moved it into netExplosion() for increased system efficiency.

rectified code formatting.